requirements.txt
Django==4.2.11
djangorestframework==3.15.0
drf-spectacular==0.27.1
dj-rest-auth[with_social]==5.0.2
django-filter==24.1
django-environ==0.11.2
psycopg[binary]==3.1.8
celery==4.4.1
psycopg[binary]